requirements.txt
Flask
sentence-transformers
scikit-learn
pandas
numpy
tqdm
gunicorn
faiss-cpu
rapidfuzz
unidecode
openpyxl
